Move error handling to RenderChainScript
Other Renderscripts might mess up the error message otherwise. For
example: when one of the SuperRes prescalers fails the error message is
unreadable.
